Clean code
