+++ config/hcaptcha.py
@@ -1,7 +1,5 @@
 from django.conf import settings
 import requests
-from dataclasses import asdict
-
 HCAPTCHA_VERIFY_URL = 'https://hcaptcha.com/siteverify'
 
 # FOR TEST: https://docs.hcaptcha.com/#test-key-set-publisher-account
@@ -19,9 +17,3 @@ def validate_hcaptcha(captcha):
 
     response_json = response.json()
     return response_json['success']
-
-
-def get_values_list_from_dataclass(data_class):
-    if data_class:
-        return asdict(data_class)
-    return {}
